requirements.txt
camelot-py==0.10.1
colorama==0.4.4
luigi==3.0.3
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pdftotext==2.2.2
pikepdf==5.1.1
readchar==3.0.5
matplotlib==3.5.1
opencv-python==4.5.5.64
